Update and rename test_send_command.py to test_send_commands.py
- add test for KeyboardInterrupt
- add verification of user output during tests
